Fixes and improvements for server fps limit Better script hook Added compiler checks(is C++11 supported, correct gcc version) Cleaned up some code

+++ Shared/SharedUtility.h
@@ -118,7 +118,7 @@ namespace SharedUtility
 
 	// TODO: TimeInit and TimeShutdown from Networked: IV
 	// 
-	unsigned long GetTime();
+	long long GetTime();
 
 	// Check if a path exists
 	bool Exists(const char * szPath);
